requirements.txt
fastapi
uvicorn
openai
faiss-cpu
sentence-transformers
redis
pymupdf
python-dotenv